+ Changed the shield shot aborption algorithm to count from zero upto a max number of shots absorbed.
+ Got rid of the main-gun, passive-gun, main-shield slots on the player class.
Those are all in the turrets now!

+ Created a file where the increase-power verb lives.

+ Fixed it so when the tesla powerup is picked up, it increases the
power of the tesla weapon again (it was broken). As a consequence,
it works for shields too and currently just resets the shots absorbed
of a shield back to zero. NOTE: I want to have this increase-power
verb with weapons too, but currently weapons are keywords in the
turret payload whereas shields and passive guns are actual instances.
I suspect I need to make a weapon an actual instance too and then
figure out what it means to increase the pwer of a weapon.

+ Removed some dead code.