Finish adding a feature where I can supply a weighted choice for something in the asset files. For example, when enemy-3 is spawned, there is a 75% chance that it will have a shot-shield and this is represented as ((.75 shot-shield)) in the turret payload location in the :port-layout for that ship.
Rectified how turret payloads are spawned/created. Now at all times
there is a real instance of an object in the payload of a turret. This
induced a change where I added a 'muzzle' object that holds the instance
name of a shot appropriate for that muzzle. This also caused a change to
SHOOT.

I'm not entirely happy with this implementation. The idea is good,
but the code needs a little bit of refactoring to make it better.