Added proximity mines for player and the fact they can be dropped and recharged. Initially you have 5 mines which can be dropped with the 'd' key.
Refactored the HUD a little bit so it will show up for mines too.
The powerup code in PERFORM-COLLIDE is a bit wonky especially when
mixed with INCREASE-POWER. It is vaguely defined  when instances get
replaced, or have their power increased, or what. I may need to redo the
powerup DSL portion a little bit in the option-9.dat file to represent
these things and possibly use REINITIALIZE-INSTANCE in addition to stuff
like INCREASE-POWER.