Implemented jutter removal using quaternion slerp interpolation.
The first time a world-basis is computed, we copy it into the previous
world basis. This ensures that while the jutter removal code might be
active, no interpolation can happen for the first rendered frame is
it happened after only one physics update. In any subsequent physics
updates, the current world-basis is copied to the previous one.