Added the beginnings of dual quaternions.
Instead of a structure containing regular quaternions to represent the
dual quaternions, which is traditionally how it is done, I represent
the dual quaternions as a single struct with real and dual components.
Consequently, this means I'll essentially copy code for handling
the dual quaternion operations. However, this ensures that I really
have a string handle on the space usage and representation for dual
quaternions. It turns out that one cannot easily type a structure
to store two quaternions in such a way as to remove any black box
chicanery concerning memory layout.

The representation may change in the future, but probably not.