Implement dual quaternion normalization. Adjust the default dual quaternion to be for a rigid body transform of 0 radians and 0 units of translation.
NOTE: This default dual quaternion is NOT normalized. Unlike regular
quaternions, it so far appears that you don't keep dual quaternions
in normalized form (which makes sense if you think about it).

Generally this is the code complete version of the dual quaternion API.
I might add some more higher order operations later, but this seems to be
the meat of what I need. Later commits in here will polish the code.