Fix spacebar and charged shot behavior.
When the spacebar is pressed it immediately emits a shot. Then, if
the shot charging powerups are in effect, if the charge timer completes
before the spacebar is let up, then a charged shot also fires.

The actual game play behavior of the charged shot prolly needs work.
Currently, the more charged shot powerups you get the faster the charge
goes until you get a shot when you hit the spacebar and then a charged
shot when you release as fast as a human can do it.

It could be that we might want the charge to just be 1 second, no matter
what, and either you have it or you don't. etc, etc.